code/api/requirements.txt
alembic==1.7.7
anyio==3.3.4
apache-libcloud==3.5.1
asgiref==3.4.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
azure-core==1.20.0
azure-cosmos==4.2.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.2.0
MarkupSafe==2.0.1
msrest==0.6.21
mysqlclient==2.1.0
oauthlib==3.1.1
orjson==3.6.4
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.16.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
